Move data product processing method
to general processing class from match specific class
