+++ src/service/match_retrieval.py
@@ -5,17 +5,16 @@ to the match, the match is in the expected state, and access times are updated c
 
 import jsonschema
 import logging
-from typing import Any, Callable
+from typing import Any
 from collections.abc import Iterable
 
-from src.service.app_state_data_structures import PickleableDependencies, CollectionsState
+from src.service.app_state_data_structures import CollectionsState
 # kinda feel like users should be more generic, but not work the trouble
 from src.service import kb_auth
 from src.service import errors
 from src.service import models
 from src.service import processing
 from src.service.matchers.common_models import Matcher
-from src.service.storage_arango import ArangoStorage
 from src.service.workspace_wrapper import WorkspaceWrapper, WORKSPACE_UPA_PATH
 
 
@@ -262,58 +261,3 @@ def _upaerror(upa, upapath, upapathlen, index):
             f"Illegal UPA '{upa}' in path '{upapath}' at index {index}")
     else:
         raise errors.IllegalParameterError(f"Illegal UPA '{upa}' at index {index}")
-
-
-# TODO NEXT move to processing.py
-async def get_or_create_data_product_process(
-    deps: CollectionsState,
-    internal_id: str,
-    data_product: str,
-    type_: models.ProcessType,
-    process_callable: Callable[[str, PickleableDependencies, list[Any]], None],
-) -> models.DataProductProcess:
-    """
-    Get a process data structure for a data product process.
-
-    Creates and starts the process the process does not already exist.
-
-    If the process exists but hasn't had a heartbeat in the required time frame, starts another
-    instance of the process.
-
-    In either case, the list of arguments to the process is empty.
-
-    deps - the collections service state.
-    internal_id - the internal ID of the parent process for the data product.
-    data_product - the data product for the process.
-    type_ - the type of the process
-    process_callable - the callable to run if there is no process information in the database or
-        the heartbeat is too old.
-    """
-    now = deps.get_epoch_ms()
-    dp_proc, exists = await deps.arangostorage.create_or_get_data_product_process(
-        models.DataProductProcess(
-            data_product=data_product,
-            type=type_,
-            internal_id=internal_id,
-            created=now,
-            state=models.ProcessState.PROCESSING,
-            state_updated=now,
-        )
-    )
-    if not exists:
-        _start_process(internal_id, process_callable, deps.get_pickleable_dependencies())
-    elif processing.requires_restart(deps.get_epoch_ms(), dp_proc):
-        logging.getLogger(__name__).warn(
-            f"Restarting {type_.value} process for internal ID {internal_id} "
-            + f"data product {data_product}"
-        )
-        _start_process(internal_id, process_callable, deps.get_pickleable_dependencies())
-    return dp_proc
-
-
-def _start_process(
-    internal_id: str,
-    process_callable: Callable[[str, PickleableDependencies, list[Any]], None],
-    deps: PickleableDependencies,
-) -> None:
-    processing.CollectionProcess(process=process_callable, args=[]).start(internal_id, deps)
