+++ src/service/processing.py
@@ -6,11 +6,12 @@ both performing the initial match and calculating secondary data products.
 import asyncio
 from apscheduler.schedulers.asyncio import AsyncIOScheduler
 from apscheduler.triggers.interval import IntervalTrigger
+import logging
 import multiprocessing
 
 from pydantic import BaseModel, Field
 from typing import Callable, Any
-from src.service.app_state_data_structures import PickleableDependencies
+from src.service.app_state_data_structures import PickleableDependencies, CollectionsState
 from src.service import models
 from src.service.timestamp import now_epoch_millis
 
@@ -130,3 +131,57 @@ class Heartbeat:
             self._schd.shutdown(wait=True)
             self._schd.remove_job(self._job_id)
             self._job_id = None
+
+
+async def get_or_create_data_product_process(
+    deps: CollectionsState,
+    internal_id: str,
+    data_product: str,
+    type_: models.ProcessType,
+    process_callable: Callable[[str, PickleableDependencies, list[Any]], None],
+) -> models.DataProductProcess:
+    """
+    Get a process data structure for a data product process.
+
+    Creates and starts the process the process does not already exist.
+
+    If the process exists but hasn't had a heartbeat in the required time frame, starts another
+    instance of the process.
+
+    In either case, the list of arguments to the process is empty.
+
+    deps - the collections service state.
+    internal_id - the internal ID of the parent process for the data product.
+    data_product - the data product for the process.
+    type_ - the type of the process
+    process_callable - the callable to run if there is no process information in the database or
+        the heartbeat is too old.
+    """
+    now = deps.get_epoch_ms()
+    dp_proc, exists = await deps.arangostorage.create_or_get_data_product_process(
+        models.DataProductProcess(
+            data_product=data_product,
+            type=type_,
+            internal_id=internal_id,
+            created=now,
+            state=models.ProcessState.PROCESSING,
+            state_updated=now,
+        )
+    )
+    if not exists:
+        _start_process(internal_id, process_callable, deps.get_pickleable_dependencies())
+    elif requires_restart(deps.get_epoch_ms(), dp_proc):
+        logging.getLogger(__name__).warn(
+            f"Restarting {type_.value} process for internal ID {internal_id} "
+            + f"data product {data_product}"
+        )
+        _start_process(internal_id, process_callable, deps.get_pickleable_dependencies())
+    return dp_proc
+
+
+def _start_process(
+    internal_id: str,
+    process_callable: Callable[[str, PickleableDependencies, list[Any]], None],
+    deps: PickleableDependencies,
+) -> None:
+    CollectionProcess(process=process_callable, args=[]).start(internal_id, deps)
